use the cleanup_stage function instead of manual cleanup for stage4, do not wipe log directories for tinderbox

+++ targets/support/chroot-functions.sh
@@ -234,7 +234,7 @@ cleanup_stages() {
 	# Remove bindist from use
 	sed -i "/USE=\"\${USE} bindist\"/d" "${clst_make_conf}"
 
-	rm -f /var/log/emerge.log /var/log/portage/elog/*
+	[ "${clst_target}" != "tinderbox" ] && rm -f /var/log/emerge.log /var/log/portage/elog/*
 }
 
 update_env_settings(){
